Refactor file sending into FileSender class
