+++ worker.py
@@ -22,6 +22,7 @@ from monitorcontrol import get_monitors
 from PySide6.QtCore import QObject, Signal, QSettings
 from hotkey_manager import HotkeyManager
 from input_streamer import InputStreamer
+from file_sender import FileSender
 from config import (
     SERVICE_TYPE,
     SERVICE_NAME_PREFIX,
@@ -68,7 +69,7 @@ class KVMWorker(QObject):
         'switch_monitor', 'local_ip', 'server_ip',
         'device_name', 'clipboard_thread', 'last_clipboard', 'server_socket',
         'network_file_clipboard', '_cancel_transfer', 'last_server_ip',
-        'hotkey_manager'
+        'hotkey_manager', 'file_sender'
     )
 
     finished = Signal()
@@ -90,6 +91,7 @@ class KVMWorker(QObject):
         self.pynput_listeners = []
         self.zeroconf = Zeroconf()
         self.input_streamer = InputStreamer(self)
+        self.file_sender = FileSender(self)
         self.switch_monitor = True
         self.local_ip = get_local_ip()
         self.server_ip = None
@@ -210,173 +212,6 @@ class KVMWorker(QObject):
             )
             return tempfile.mkdtemp()
 
-    def _create_archive(self, paths, cancel_event: Optional[threading.Event] = None):
-        temp_dir = self._get_temp_dir()
-        archive = os.path.join(temp_dir, 'share.zip')
-        logging.debug("Created temp archive dir %s", temp_dir)
-        # Log available space in the temporary directory which will hold the archive
-        try:
-            usage = shutil.disk_usage(temp_dir)
-            logging.debug(
-                "Temp dir disk usage - total: %s, used: %s, free: %s",
-                usage.total,
-                usage.used,
-                usage.free,
-            )
-        except Exception as e:
-            logging.debug("Failed to query temp dir disk usage: %s", e)
-        start_time = time.time()
-        try:
-            # Pre-scan all paths to determine total number of files
-            total_files = 0
-            for p in paths:
-                if os.path.isdir(p):
-                    for _, _, files in os.walk(p):
-                        total_files += len(files)
-                else:
-                    total_files += 1
-
-            archived_files = 0
-            with zipfile.ZipFile(archive, 'w', zipfile.ZIP_DEFLATED, allowZip64=True) as zf:
-                def _write_with_progress(src_path, arcname, ctype):
-                    file_size = os.path.getsize(src_path)
-                    if cancel_event and cancel_event.is_set():
-                        raise RuntimeError('archive canceled')
-
-                    last_percentage = -1
-                    last_emit_time = time.time()
-                    start_time = time.time()
-
-                    if file_size > 1_000_000_000:
-                        info = zipfile.ZipInfo(arcname, date_time=time.localtime(time.time())[:6])
-                        info.compress_type = ctype
-                        with open(src_path, 'rb') as src_file, zf.open(info, 'w', force_zip64=True) as dest:
-                            sent = 0
-                            while True:
-                                if cancel_event and cancel_event.is_set():
-                                    raise RuntimeError('archive canceled')
-                                chunk = src_file.read(FILE_CHUNK_SIZE)
-                                if not chunk:
-                                    break
-                                dest.write(chunk)
-                                sent += len(chunk)
-
-                                current_percentage = int((sent / file_size) * 100)
-                                if current_percentage > last_percentage or time.time() - last_emit_time > PROGRESS_UPDATE_INTERVAL:
-                                    # --- Speed and ETR Calculation ---
-                                    elapsed_time = time.time() - start_time
-                                    speed_mbps = (sent / (1024*1024)) / elapsed_time if elapsed_time > 0 else 0
-                                    remaining_bytes = file_size - sent
-                                    etr_seconds = int(remaining_bytes / (speed_mbps * 1024 * 1024)) if speed_mbps > 0 else 0
-                                    etr_str = time.strftime('%M:%S', time.gmtime(etr_seconds)) if etr_seconds < 3600 else time.strftime('%H:%M:%S', time.gmtime(etr_seconds))
-
-                                    label = f"{os.path.basename(src_path)}: {sent/1024/1024:.1f}MB / {file_size/1024/1024:.1f}MB\n"
-                                    label += f"Sebesség: {speed_mbps:.1f} MB/s | Hátralévő idő: {etr_str}"
-
-                                    self.update_progress_display.emit(current_percentage, label)
-                                    last_percentage = current_percentage
-                                    last_emit_time = time.time()
-                        final_label = f"{os.path.basename(src_path)}: Kész! ({file_size/1024/1024:.1f}MB)"
-                        self.update_progress_display.emit(100, final_label)
-                    else:
-                        zf.write(src_path, arcname, compress_type=ctype)
-                for p in paths:
-                    if os.path.isdir(p):
-                        base = os.path.basename(p.rstrip(os.sep))
-                        for root, _, files in os.walk(p):
-                            for f in files:
-                                full = os.path.join(root, f)
-                                rel = os.path.join(base, os.path.relpath(full, p))
-                                file_size = os.path.getsize(full)
-                                if file_size > 1_000_000_000:
-                                    logging.info(
-                                        "Adding large file %s (%d bytes) to archive",
-                                        full,
-                                        file_size,
-                                    )
-                                try:
-                                    compress_type = zipfile.ZIP_DEFLATED
-                                    ext = os.path.splitext(f)[1].lower()
-                                    if file_size > 1_000_000_000 or ext in {'.mkv', '.mp4', '.mov'}:
-                                        compress_type = zipfile.ZIP_STORED
-                                    _write_with_progress(full, rel, compress_type)
-                                    logging.debug("Archived %s", full)
-                                    if cancel_event and cancel_event.is_set():
-                                        raise RuntimeError('archive canceled')
-                                except MemoryError:
-                                    msg = (
-                                        f"Archiv\xe1l\xe1si hiba: Kev\xe9s a mem\xf3ria a(z) {os.path.basename(full)} t\xf6m\xf6r\xedt\xe9s\xe9hez."
-                                    )
-                                    logging.error(msg, exc_info=True)
-                                    self.file_transfer_error.emit(msg)
-                                    return None
-                                except (IOError, OSError) as e:
-                                    msg = f"Archiv\xe1l\xe1si hiba: Nincs el\xe9g hely vagy IO probl\xe9ma ({e})."
-                                    logging.error(msg, exc_info=True)
-                                    self.file_transfer_error.emit(msg)
-                                    return None
-                                except zipfile.LargeZipFile as e:
-                                    msg = f"Archiv\xe1l\xe1si hiba: {e}"
-                                    logging.error(msg, exc_info=True)
-                                    self.file_transfer_error.emit(msg)
-                                    return None
-                                archived_files += 1
-                                percentage = int(archived_files / total_files * 100) if total_files else 0
-                                label = f"Tömörítés: {os.path.basename(full)} ({archived_files}/{total_files} fájl)"
-                                self.update_progress_display.emit(percentage, label)
-                    else:
-                        file_size = os.path.getsize(p)
-                        if file_size > 1_000_000_000:
-                            logging.info(
-                                "Adding large file %s (%d bytes) to archive",
-                                p,
-                                file_size,
-                            )
-                        try:
-                            compress_type = zipfile.ZIP_DEFLATED
-                            ext = os.path.splitext(p)[1].lower()
-                            if file_size > 1_000_000_000 or ext in {'.mkv', '.mp4', '.mov'}:
-                                compress_type = zipfile.ZIP_STORED
-                            _write_with_progress(p, os.path.basename(p), compress_type)
-                            logging.debug("Archived %s", p)
-                            if cancel_event and cancel_event.is_set():
-                                raise RuntimeError('archive canceled')
-                        except MemoryError:
-                            msg = (
-                                f"Archiv\xe1l\xe1si hiba: Kev\xe9s a mem\xf3ria a(z) {os.path.basename(p)} t\xf6m\xf6r\xedt\xe9s\xe9hez."
-                            )
-                            logging.error(msg, exc_info=True)
-                            self.file_transfer_error.emit(msg)
-                            return None
-                        except (IOError, OSError) as e:
-                            msg = f"Archiv\xe1l\xe1si hiba: Nincs el\xe9g hely vagy IO probl\xe9ma ({e})."
-                            logging.error(msg, exc_info=True)
-                            self.file_transfer_error.emit(msg)
-                            return None
-                        except zipfile.LargeZipFile as e:
-                            msg = f"Archiv\xe1l\xe1si hiba: {e}"
-                            logging.error(msg, exc_info=True)
-                            self.file_transfer_error.emit(msg)
-                            return None
-                        archived_files += 1
-                        percentage = int(archived_files / total_files * 100) if total_files else 0
-                        label = f"Tömörítés: {os.path.basename(p)} ({archived_files}/{total_files} fájl)"
-                        self.update_progress_display.emit(percentage, label)
-            label = f"Tömörítés kész. ({os.path.basename(archive)})"
-            self.update_progress_display.emit(100, label)
-        except Exception as e:
-            logging.error("Failed to create archive: %s", e, exc_info=True)
-            try:
-                shutil.rmtree(temp_dir, ignore_errors=True)
-            except Exception as cleanup_err:
-                logging.error("Failed to cleanup temp dir %s: %s", temp_dir, cleanup_err)
-            self.file_transfer_error.emit(f"Archive creation failed: {e}")
-            return None
-        duration = time.time() - start_time
-        if duration > 600:
-            logging.warning("Archive creation took %.1f seconds", duration)
-        logging.debug("Archive created at %s", archive)
-        return archive
 
     def _safe_extract_archive(self, archive_path, dest_dir):
         """Extract archive to dest_dir and cleanup on failure."""
@@ -419,79 +254,6 @@ class KVMWorker(QObject):
             shutil.rmtree(temp_extract, ignore_errors=True)
             logging.debug("Extraction complete, removed %s", temp_extract)
 
-    def _send_archive(self, sock, archive_path, dest_dir):
-        self._cancel_transfer.clear()
-        logging.debug(
-            "Entering _send_archive. cancel=%s dest=%s",
-            self._cancel_transfer.is_set(),
-            dest_dir,
-        )
-        logging.debug("Cancel flag cleared at start of _send_archive")
-        prev_to = sock.gettimeout()
-        sock.settimeout(TRANSFER_TIMEOUT)
-        try:
-            size = os.path.getsize(archive_path)
-            name = os.path.basename(archive_path)
-            meta = {
-                'type': 'file_metadata',
-                'name': name,
-                'size': size,
-                'dest': dest_dir,
-                'source_id': self.network_file_clipboard.get('source_id') if self.network_file_clipboard else self.device_name,
-            }
-            if not self._send_message(sock, meta):
-                return
-            sent = 0
-            last_percentage = -1
-            last_emit_time = time.time()
-            start_time = time.time()  # For average speed calculation
-
-            with open(archive_path, 'rb') as f:
-                while not self._cancel_transfer.is_set():
-                    chunk = f.read(FILE_CHUNK_SIZE)
-                    if not chunk:
-                        break
-                    if not self._send_message(sock, {'type': 'file_chunk', 'data': chunk}):
-                        raise IOError('send failed')
-                    sent += len(chunk)
-
-                    # Throttle the UI update
-                    if time.time() - last_emit_time >= PROGRESS_UPDATE_INTERVAL:
-                        current_percentage = int((sent / size) * 100) if size > 0 else 0
-
-                        # --- Speed and ETR Calculation ---
-                        elapsed_time = time.time() - start_time
-                        speed_mbps = (sent / (1024*1024)) / elapsed_time if elapsed_time > 0 else 0
-                        remaining_bytes = size - sent
-                        etr_seconds = int(remaining_bytes / (speed_mbps * 1024 * 1024)) if speed_mbps > 0 else 0
-                        etr_str = time.strftime('%M:%S', time.gmtime(etr_seconds)) if etr_seconds < 3600 else time.strftime('%H:%M:%S', time.gmtime(etr_seconds))
-
-                        label = f"{name}: {sent/1024/1024:.1f}MB / {size/1024/1024:.1f}MB\n"
-                        label += f"Sebesség: {speed_mbps:.1f} MB/s | Hátralévő idő: {etr_str}"
-
-                        self.update_progress_display.emit(current_percentage, label)
-                        last_emit_time = time.time()
-            if self._cancel_transfer.is_set():
-                self._send_message(sock, {'type': 'transfer_canceled'})
-                return
-            final_label = f"{name}: Kész! ({size/1024/1024:.1f}MB)"
-            self.update_progress_display.emit(100, final_label)
-            self._send_message(sock, {'type': 'file_end'})
-            logging.debug(
-                "WORKER EMITTING update_progress_display: %s %d/%d",
-                name,
-                size,
-                size,
-            )
-            logging.debug("_send_archive loop completed. cancel=%s", self._cancel_transfer.is_set())
-        except Exception as e:
-            logging.error('Error sending archive: %s', e, exc_info=True)
-            self.file_transfer_error.emit(str(e))
-        finally:
-            sock.settimeout(prev_to)
-            self._cancel_transfer.clear()
-            logging.debug("Archive send finished, cancel flag cleared")
-            logging.debug("Exiting _send_archive")
 
     def _clear_network_file_clipboard(self):
         """Remove any stored temporary archive and clear the clipboard info."""
@@ -530,13 +292,21 @@ class KVMWorker(QObject):
             self._cancel_transfer.is_set(),
         )
         logging.debug("Cancel flag cleared at start of _share_files_thread")
+        if self.settings['role'] != 'ado':
+            sock = self.server_socket
+            if not sock:
+                logging.warning('No server connection for file share')
+                self.file_transfer_error.emit("Nincs kapcsolat a szerverrel a küldéshez.")
+                return
+            self.file_sender.send_files(paths, operation, sock)
+            return
         timeout = self.settings.get('archive_timeout_seconds', 900)
         cancel_evt = threading.Event()
         result = {}
         temp_archive_dir = None
 
         def run_archiving():
-            result['archive'] = self._create_archive(paths, cancel_event=cancel_evt)
+            result['archive'] = self.file_sender._create_archive(paths, cancel_event=cancel_evt)
 
         arch_thread = threading.Thread(target=run_archiving, daemon=True)
         arch_thread.start()
@@ -577,17 +347,8 @@ class KVMWorker(QObject):
                     'source_id': self.device_name,
                     'operation': operation,
                 })
-            else:  # This is the client-side sending logic
-                sock = self.server_socket
-                if not sock:
-                    logging.warning('No server connection for file share')
-                    self.file_transfer_error.emit("Nincs kapcsolat a szerverrel a küldéshez.")
-                    return
-
-                # Use the robust _send_archive method for sending.
-                # The 'dest' parameter for _send_archive is not used on the sending side,
-                # but the method expects it. We can pass an empty string.
-                self._send_archive(sock, archive, dest_dir="")
+            else:
+                pass  # client-side handled earlier
         finally:
             if temp_archive_dir and self.settings['role'] != 'ado':
                 logging.info(f"Client-side cleanup: Removing temporary archive directory {temp_archive_dir}")
@@ -874,7 +635,7 @@ class KVMWorker(QObject):
                                 if self.network_file_clipboard and self.network_file_clipboard.get('archive'):
                                     self._cancel_transfer.clear()
                                     logging.debug("Cancel flag cleared for paste_request")
-                                    self._send_archive(sock, self.network_file_clipboard['archive'], dest)
+                                    self.file_sender._send_archive(sock, self.network_file_clipboard['archive'], dest)
                             elif data.get('type') == 'file_metadata':
                                 logging.info("[WORKER_DEBUG] Received 'upload_file_start' from client: %s (size: %s)", data.get('name'), data.get('size'))
                                 temp_dir_for_download = self._get_temp_dir()
